♻️ Refactor : add updateDates
